+++ factory/src/test/java/io/github/factoryfx/factory/BranchCreatorTest.java
@@ -7,6 +7,7 @@ import io.github.factoryfx.factory.testfactories.ExampleFactoryA;
 import io.github.factoryfx.factory.testfactories.ExampleFactoryB;
 import io.github.factoryfx.factory.testfactories.ExampleLiveObjectA;
 import io.github.factoryfx.factory.testfactories.ExampleLiveObjectB;
+import org.checkerframework.checker.units.qual.K;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.mockito.internal.util.MockUtil;
@@ -81,10 +82,33 @@ class BranchSelectorTest {
 
         BranchSelector<ExampleFactoryA> branchSelector = new BranchSelector<>(root);
         ExampleLiveObjectB mock = Mockito.mock(ExampleLiveObjectB.class);
-        branchSelector.select(ExampleFactoryB.class).mock(mock);
+        branchSelector.select(ExampleFactoryB.class,"bla").mock(mock);
         assertEquals(mock,root.referenceAttribute.get().internal().instance());
         assertFalse(MockUtil.isMock(root.internal().instance()));
     }
+
+
+    @Test
+    public void test_name_null(){
+        FactoryTreeBuilder<ExampleLiveObjectA,ExampleFactoryA,Void> builder = new FactoryTreeBuilder<>(ExampleFactoryA.class, ctx -> {
+            ExampleFactoryA root = new ExampleFactoryA();
+            root.referenceListAttribute.add(ctx.get(ExampleFactoryB.class,"1"));
+            root.referenceListAttribute.add(ctx.get(ExampleFactoryB.class,"2"));
+            root.referenceListAttribute.add(ctx.get(ExampleFactoryB.class));
+            return root;
+        });
+        builder.addFactory(ExampleFactoryB.class, "1",Scope.SINGLETON, ctx -> new ExampleFactoryB());
+        builder.addFactory(ExampleFactoryB.class, "2",Scope.SINGLETON, ctx -> new ExampleFactoryB());
+        builder.addFactory(ExampleFactoryB.class, Scope.SINGLETON, ctx -> {
+            ExampleFactoryB exampleFactoryB = new ExampleFactoryB();
+            exampleFactoryB.stringAttribute.set("bla");
+            return exampleFactoryB;
+        });
+
+        BranchSelector<ExampleFactoryA> subTreeUtility = new BranchSelector<>(builder);
+        assertEquals("bla",subTreeUtility.select(ExampleFactoryB.class).factory().stringAttribute.get());
+    }
+
 
 
 }